Ecommerce y mejoras en las interfaces.

+++ app/Console/Commands/install.php
@@ -13,14 +13,15 @@ class install extends Command
      *
      * @var string
      */
-    protected $signature = 'template:install';
+    protected $signature = 'ecommerce:install
+                            {--r|reset : Reset database}';
 
     /**
      * The console command description.
      *
      * @var string
      */
-    protected $description = 'Install Laravel template';
+    protected $description = 'Install Laravel Ecommerce';
 
     /**
      * Create a new command instance.
@@ -39,11 +40,29 @@ class install extends Command
      */
     public function handle()
     {
-        $this->call('key:generate');
-        $this->call('migrate:fresh');
-        $this->call('db:seed');
-        $this->call('storage:link');
-        $this->call('vendor:publish', ['--provider' => VoyagerServiceProvider::class, '--tag' => ['config', 'voyager_avatar']]);
-        $this->info('Gracias por instalar LaravelTemplate');
+        if($this->option('reset')){
+            $this->call('migrate:fresh');
+            $this->call('db:seed');
+            $this->info('La base de datos de Laravel Ecommerce ha sido reiniciada ;)');
+        }else{
+            
+            $empty_database = false;
+            try {
+                DB::table('users')->get();
+            } catch (\Throwable $th) {
+                $empty_database = true;
+            }
+
+            if(!$empty_database){
+                $this->call('key:generate');
+                $this->call('migrate');
+                $this->call('db:seed');
+                $this->call('storage:link');
+                $this->call('vendor:publish', ['--provider' => VoyagerServiceProvider::class, '--tag' => ['config', 'voyager_avatar']]);
+                $this->info('Gracias por instalar Laravel Ecommerce!!!');
+            }else{
+                $this->error('La base de datos de Laravel Ecommerce ya está instalada.');
+            }
+        }
     }
 }
